+++ wilfred/keyboard.py
@@ -1,18 +1,19 @@
-#################################################################
-#                                                               #
-# Wilfred                                                       #
-# Copyright (C) 2020, Vilhelm Prytz, <vilhelm@prytznet.se>      #
-#                                                               #
-# Licensed under the terms of the MIT license, see LICENSE.     #
-# https://github.com/wilfred-dev/wilfred                        #
-#                                                               #
-#################################################################
+####################################################################
+#                                                                  #
+# Wilfred                                                          #
+# Copyright (C) 2020, Vilhelm Prytz, <vilhelm@prytznet.se>, et al. #
+#                                                                  #
+# Licensed under the terms of the MIT license, see LICENSE.        #
+# https://github.com/wilfred-dev/wilfred                           #
+#                                                                  #
+####################################################################
 
 import threading
 
 
 class KeyboardThread(threading.Thread):
     def __init__(self, input_callback, params):
+        self._running = True
         self.input_callback = input_callback
         self.params = params
 
@@ -21,5 +22,8 @@ class KeyboardThread(threading.Thread):
         self.start()
 
     def run(self):
-        while True:
-            self.input_callback(input(), self.params)
+        while self._running:
+            try:
+                self.input_callback(input(), self.params)
+            except Exception:
+                self._running = False